utils/Mitigation Prioritization In a Caldera Operation/requirements.txt
requests==2.31.0
neo4j==5.12.0
pandas==2.2.2
pyyaml==6.0
numpy==1.26.0
tabulate==0.9.0
